Change params to reflect Cropster naming
Cropster users have both a username/password and an api key/secret. When
using the api, we need to authorize using the api key/secret. This
changes the params the client expects to match the naming that Cropster
uses and reduce confusion around authorization.

This is a major version bump because all clients will need to update
their integration to use these changes.